[Refactor] Point: point, history request 패키지 구분 #155
