+++ point/src/main/java/com/haot/point/application/dto/request/point/PointCreateRequest.java
@@ -1,4 +1,4 @@
-package com.haot.point.application.dto.request;
+package com.haot.point.application.dto.request.point;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
