+++ point/src/main/java/com/haot/point/application/dto/request/point/PointUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.haot.point.application.dto.request;
+package com.haot.point.application.dto.request.point;
 
 import jakarta.validation.constraints.Min;
