+++ point/src/main/java/com/haot/point/application/service/PointService.java
@@ -1,6 +1,6 @@
 package com.haot.point.application.service;
 
-import com.haot.point.application.dto.request.PointCreateRequest;
+import com.haot.point.application.dto.request.point.PointCreateRequest;
 import com.haot.point.application.dto.request.PointTransactionRequest;
 import com.haot.point.application.dto.response.PointAllResponse;
 import com.haot.point.application.dto.response.PointResponse;
