+++ point/src/main/java/com/haot/point/presentation/controller/PointHistoryController.java
@@ -1,6 +1,6 @@
 package com.haot.point.presentation.controller;
 
-import com.haot.point.application.dto.request.PointStatusRequest;
+import com.haot.point.application.dto.request.point.PointStatusRequest;
 import com.haot.point.application.dto.response.PointAllResponse;
 import com.haot.point.application.dto.response.PointHistoryResponse;
 import com.haot.point.application.service.PointHistoryService;
